Moved jobs to DesignatedMeter class to create endpoints for each job
